gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyJWT==2.8.0
pydantic-settings==2.0.3
python-dotenv == 1.0.0

# async http requests & fast dns resolving
aiohttp==3.8.6
aiodns==3.1.1

# for test purposes
flake8==6.1.0
ipdb==0.13.13
ipython==8.17.2